Clean up all pre-2021 crap
