+++ rabbitmq_client/__init__.py
@@ -1,7 +1,7 @@
 import logging
 
-from rabbitmq_client.new_connection import RMQConnection  # noqa
-from rabbitmq_client.new_consumer import RMQConsumer  # noqa
+from rabbitmq_client.connection import RMQConnection  # noqa
+from rabbitmq_client.consumer import RMQConsumer  # noqa
 from rabbitmq_client.defs import (  # noqa
     QueueParams,
     ExchangeParams,
