+++ rabbitmq_client/consumer.py
@@ -1,297 +1,405 @@
+import functools
 import logging
 
-from threading import Thread
-
-from rabbitmq_client.errors import ConsumerAlreadyExists
-
-from rabbitmq_client.consumer_connection import RMQConsumerConnection
-from rabbitmq_client.common_defs import Printable
-from rabbitmq_client.consumer_defs import (
-    Subscription,
-    ConsumedMessage,
-    ConsumeOk,
-    RPCServer,
-    RPCClient,
-    CommandQueue
-)
+from rabbitmq_client.defs import QueueParams, QueueBindParams, ConsumeOK
+from rabbitmq_client.connection import RMQConnection
 
 
 LOGGER = logging.getLogger(__name__)
 
 
-class Consumer(Printable):
+def _gen_consume_key(queue=None,
+                     exchange=None,
+                     routing_key=None):
     """
-    Used to package what makes up a consumer to the RMQConsumer class. This
-    class is instantiated when a new subscription is made.
+    :param queue: str
+    :param exchange: str
+    :param routing_key: str
+    :return: str
+    """
+    key_list = []
+    key_list.append(queue) if queue else ""
+    key_list.append(exchange) if exchange else ""
+    key_list.append(routing_key) if routing_key else ""
 
-    The preserve flag is used to subscribe internally in the RMQ client, where
-    the raw message format (ConsumedMessage) is used instead of extracting only
-    the delivered message.
+    separator = "|"
+
+    return separator.join(key_list)
+
+
+class RMQConsume:
+    """
+    Used by the RMQConsumer to keep track of which consumes have been started
+    so far, and to be able to re-distribute consumes should the underlying
+    connection experience a hiccup.
     """
 
     def __init__(self,
-                 callback,
-                 preserve=False,
-                 exchange="",
-                 routing_key="",
-                 consumer_tag=""):
+                 consume_params,
+                 queue_params,
+                 exchange_params,
+                 routing_key):
         """
-        :param callback: callback to call on new message
-        :param preserve: set if you want to preserve original message format:
-                         ConsumedMessage instead of only getting bytes.
-        :param exchange: subscribed exchange
-        :param routing_key: subscribed routing key (queue name in case of
-                            default exchange)
+        :param consume_params: rabbitmq_client.ConsumeParams
+        :param queue_params: rabbitmq_client.QueueParams
+        :param exchange_params: rabbitmq_client.ExchangeParams
+        :param routing_key: str
         """
-        LOGGER.debug("__init__ Consumer")
-
-        self.callback = callback
-        self.preserve = preserve
-        self.exchange = exchange
+        self.consume_params = consume_params
+        self.queue_params = queue_params
+        self.exchange_params = exchange_params
         self.routing_key = routing_key
-        self.consumer_tag = consumer_tag
 
-    def set_consumer_tag(self, new_consumer_tag):
-        """
-        Setter for consumer_tag
+        # The actual consumer tag from 'basic_consume'. This will be equal to
+        # 'self.consume_params.consumer_tag' if one was set by the caller,
+        # otherwise it is randomly generated by pika. It is used to cancel
+        # consumes with 'basic_cancel'.
+        self._consumer_tag = None
 
-        :param new_consumer_tag: new consumer_tag
-        :return: new consumer_tag
-        """
-        LOGGER.debug(f"setting consumer tag to: {new_consumer_tag}")
+    @property
+    def consumer_tag(self):
+        return self._consumer_tag
 
-        self.consumer_tag = new_consumer_tag
-        return self.consumer_tag
+    @consumer_tag.setter
+    def consumer_tag(self, new_value):
+        self._consumer_tag = new_value
 
 
-class RMQConsumer:
+class RMQConsumer(RMQConnection):
     """
-    Implements a consumer for RabbitMQ.
+    Generic consumer implementation using the RMQConnection base class to ease
+    connection and channel handling.
+
+    The 'consume' method provided by this consumer implementation accepts
+    parameters as they will be passed to the pika library to avoid going out
+    of its way to create another layer of special handling. A few parameters
+    have been removed, but nothing has been added.
     """
 
     def __init__(self, connection_parameters=None):
         """
-        :param connection_parameters: connection parameters to the RMQ server
-        :type connection_parameters: pika.ConnectionParameters
+        :param connection_parameters: pika.ConnectionParameters
         """
-        LOGGER.debug("__init__")
+        super().__init__(connection_parameters=connection_parameters)
 
-        self._consumers = list()
+        self._ready = False
+        self._consumes = dict()
 
-        self._connection = RMQConsumerConnection(
-            self._on_message,
-            connection_parameters=connection_parameters
-        )
+    @property
+    def ready(self):
+        """
+        Indicates if the consumer is ready, meaning it will immediately issue
+        consume-work it receives. If the consumer is NOT ready, incoming
+        consumes will be delayed until the underlying connection reports ready
+        through the 'on_ready' hook.
 
-        self._connection_thread = Thread(target=self.connect)
+        :return: bool
+        """
+        return self._ready
 
     def start(self):
         """
-        Starts the RMQConsumer, meaning it is prepared for consuming messages.
+        Starts the consumer by initiating the underlying connection.
 
-        By starting the RMQConsumer, a process is created which will hold an
-        RMQConsumerConnection. This function also starts a thread in the
-        current process that monitors the consumed_messages queue for incoming
-        messages.
+        NOTE! This is NOT a synchronous operation! If you must know that the
+        consumer is successfully started, monitor the 'ready' property.
         """
-        LOGGER.info("start")
+        LOGGER.info("starting consumer")
 
-        self._connection_thread.start()
+        super().start()
+
+    def restart(self):
+        """Restarts the consumer by re-initiating the underlying connection."""
+        LOGGER.info("restarting consumer")
+
+        super().restart()
 
     def stop(self):
+        """Stops the consumer by stopping the underlying connection."""
+        LOGGER.info("stopping consumer")
+
+        super().stop()
+
+    def consume(self,
+                consume_params,
+                queue_params=None,
+                exchange_params=None,
+                routing_key=None):
         """
-        Stops the RMQConsumer, tearing down the RMQConsumerConnection process.
+        General consumer interface, when wanting to consumer messages sent to a
+        specific queue or exchange. Input parameters are a subset of those used
+        by pika, to avoid re-inventing the wheel with special abstractions.
+        The only parameters that have been removed are the "passive" and
+        "callback" options.
+
+        Some restrictions apply:
+        1. EITHER queue OR exchange OR both MUST be set.
+        2. Queue/exchange parameter change is not supported. For example, this
+           consumer will not handle a re-declaration of a queue with new
+           provided parameters.
+
+        RMQConsumer will automatically re-issue queue and exchange declarations
+        and bindings if the connection is for some reason closed and re-opened
+        at any point.
+
+        This method may be called before the underlying RMQConnection has
+        successfully established a connection and channel to RabbitMQ. Those
+        buffered actions called before RMQConnection reports it is ready,
+        through the 'on_ready' hook, will be issued immediately upon 'on_ready'
+        being called. Monitor the RMQConsumer 'ready' property to keep track of
+        when the RMQConnection has reported it is ready.
+
+        :param consume_params: rabbitmq_client.ConsumeParams
+        :param queue_params: rabbitmq_client.QueueParams
+        :param exchange_params: rabbitmq_client.ExchangeParams
+        :param routing_key: str
+        :returns: str
         """
-        LOGGER.info("stop")
+        LOGGER.info("starting consume")
 
-        self._connection.disconnect()
-        self._connection_thread.join()
+        # 1. Checks
+        if queue_params is None and exchange_params is None:
+            raise ValueError(
+                "You need to provide either a queue, and exchange, or both, "
+                "else there is nothing to consume from..."
+            )
 
-    def add_new_consumer(self, new_consumer: Consumer):
+        consume_key = _gen_consume_key(
+            queue_params.queue if queue_params else "",
+            exchange_params.exchange if exchange_params else "",
+            routing_key
+        )
+        if self._consumes.get(consume_key) is not None:
+            raise ValueError(
+                "That combination of queue + exchange + routing key already "
+                "exists."
+            )
+
+        # 2. Update consumer instance
+        self._consumes[consume_key] = RMQConsume(
+            consume_params, queue_params, exchange_params, routing_key
+        )
+
+        # 3. Start declaring shit
+        if self.ready:
+            self._handle_consume(consume_params,
+                                 queue_params,
+                                 exchange_params,
+                                 routing_key)
+
+        return consume_key
+
+    def _handle_consume(self,
+                        consume_params,
+                        queue_params,
+                        exchange_params,
+                        routing_key):
         """
-        Adds a new Consumer instance to the RMQConsumer's list of consumers, as
-        long as the new consumer is not clashing with an old one.
-
-        :param new_consumer: new consumer instance to add
-        :raises ConsumerAlreadyExists: raised if the supplied consumer clashes
-                                       with an existing one
+        :param consume_params: rabbitmq_client.ConsumeParams
+        :param queue_params: None | rabbitmq_client.QueueParams
+        :param exchange_params: None | rabbitmq_client.ExchangeParams
+        :param routing_key: None | str
         """
-        matching_consumers = \
-            [consumer for consumer in self._consumers
-             if consumer.exchange == new_consumer.exchange and
-             consumer.routing_key == new_consumer.routing_key]
+        if queue_params is None:
+            queue_params = QueueParams("", exclusive=True)
 
-        if matching_consumers:
-            raise ConsumerAlreadyExists("A matching consumer already exists")
+        cb = functools.partial(self.on_queue_declared,
+                               consume_params=consume_params,
+                               queue_params=queue_params,
+                               exchange_params=exchange_params,
+                               routing_key=routing_key)
 
-        self._consumers.append(new_consumer)
+        self.declare_queue(queue_params, callback=cb)
 
-    def connect(self):
+    def on_queue_declared(self,
+                          frame,
+                          consume_params=None,
+                          queue_params=None,
+                          exchange_params=None,
+                          routing_key=None):
         """
-        Initiates the underlying connection.
+        :param frame: pika.frame.Method
+        :param consume_params: rabbitmq_client.ConsumeParams
+        :param queue_params: rabbitmq_client.QueueParams
+        :param exchange_params: rabbitmq_client.ExchangeParams
+        :param routing_key: str
         """
-        self._connection.connect()
-        print("stopping consumer, ioloop broken")
+        LOGGER.info(f"declared queue: {frame.method.queue}")
 
-    def _on_message(self, message):
+        # Update the consume queue name to ensure it is set to the created
+        # queue's name
+        consume_params.queue = frame.method.queue
+
+        if exchange_params is not None:
+            cb = functools.partial(self.on_exchange_declared,
+                                   consume_params=consume_params,
+                                   queue_params=queue_params,
+                                   exchange_params=exchange_params,
+                                   routing_key=routing_key)
+
+            self.declare_exchange(exchange_params, callback=cb)
+
+        else:
+            cb = functools.partial(self.on_consume_ok,
+                                   queue_params=queue_params)
+
+            self.consume_from_queue(consume_params,
+                                    on_message_callback_override=self.on_msg,
+                                    callback=cb)
+
+    def on_exchange_declared(self,
+                             frame,
+                             consume_params=None,
+                             queue_params=None,
+                             routing_key=None):
         """
+        :param frame: pika.frame.Method
+        :param consume_params: rabbitmq_client.ConsumeParams
+        :param queue_params: rabbitmq_client.QueueParams
+        :param routing_key: str
         """
-        LOGGER.debug("got new consumed message")
+        LOGGER.info(f"declared exchange: {frame.method.exchange}")
 
-        # From Consumer connection
-        if isinstance(message, ConsumedMessage):
-            self.handle_message(message)
-        elif isinstance(message, ConsumeOk):
-            self.handle_consume_ok(message)
+        cb = functools.partial(self.on_queue_bound,
+                               consume_params=consume_params,
+                               queue_params=queue_params,
+                               exchange=frame.method.exchange,
+                               routing_key=routing_key)
 
-    def handle_message(self, message: ConsumedMessage):
+        self.bind_queue(QueueBindParams(consume_params.queue,
+                                        frame.method.exchange,
+                                        routing_key=routing_key),
+                        callback=cb)
+
+    def on_queue_bound(self,
+                       frame,
+                       consume_params=None,
+                       queue_params=None,
+                       exchange=None,
+                       routing_key=None):
         """
-        Defines handling for a received message, dispatches the message
-        contents to registered callbacks depending on the topic.
-
-        :param ConsumedMessage message: received message
+        :param frame: pika.frame.Method
+        :param consume_params: rabbitmq_client.ConsumeParams
+        :param queue_params: rabbitmq_client.QueueParams
+        :param exchange: str
+        :param routing_key: str
         """
-        LOGGER.debug("handle_message message: {}".format(message))
+        LOGGER.info(f"queue {consume_params.queue} bound to exchange "
+                    f"{exchange}, frame={frame}")
 
-        for consumer in self._consumers:
-            if message.exchange == consumer.exchange and \
-               message.routing_key == consumer.routing_key:
-                if consumer.preserve:  # Internal consumer, forward object
-                    consumer.callback(message)
-                else:  # Only message content
-                    consumer.callback(message.message)
+        cb = functools.partial(self.on_consume_ok,
+                               queue_params=queue_params,
+                               exchange=exchange,
+                               routing_key=routing_key)
 
-    def handle_consume_ok(self, message: ConsumeOk):
+        self.consume_from_queue(consume_params,
+                                on_message_callback_override=self.on_msg,
+                                callback=cb)
+
+    def on_consume_ok(self,
+                      frame,
+                      queue_params=None,
+                      exchange=None,
+                      routing_key=None):
         """
-        Defines handling for a received ConsumeOk message, meaning a consume
-        has been successfully started for a consumer.
-
-        :param message: contains information about the started consumer
+        :param frame: pika.frame.Method
+        :param queue_params: rabbitmq_client.QueueParams
+        :param exchange: str
+        :param routing_key: str
         """
-        LOGGER.info(f"handle_consume_ok for: {message.consume.queue_name}")
+        LOGGER.info(f"consume OK for queue: {frame}")
 
-        if isinstance(message.consume, Subscription):
-            # filter out subscription consumer and set consumer_tag
-            [consumer.set_consumer_tag(message.consumer_tag)
-             for consumer in self._consumers
-             if consumer.exchange == message.consume.topic]
-        else:  # RPCServer or RPCClient
-            [consumer.set_consumer_tag(message.consumer_tag)
-             for consumer in self._consumers
-             if consumer.routing_key == message.consume.queue_name]
+        consume_instance = self._consumes[
+            _gen_consume_key(queue=queue_params.queue,
+                             exchange=exchange,
+                             routing_key=routing_key)
+        ]
+        # Update with real consumer tag, may or may not be the same tag.
+        consume_instance.consumer_tag = frame.method.consumer_tag
 
-        consumer_state = ""
-        for consumer in self._consumers:
-            consumer_state = consumer_state + "\n\t{}".format(consumer)
+        # Enables lookup via consumer tag in 'on_msg'. These entries are
+        # removed 'on_close' since the consumer tags may be refreshed on
+        # reconnecting.
+        self._consumes[consume_instance.consumer_tag] = consume_instance
 
-        LOGGER.debug("handle_consume_ok new consumer state: {}"
-                     .format(consumer_state))
+        try:
+            consume_instance.consume_params.on_message_callback(
+                ConsumeOK(consume_instance.consumer_tag)
+            )
+        except Exception as e:
+            LOGGER.critical(f"sending consume OK to message callback resulted "
+                            f"in an exception: {e}")
 
-    def subscribe(self, topic, callback):
+    def on_msg(self, channel, basic_deliver, _basic_properties, body):
         """
-        Subscribes to messages sent to the named topic. Messages received on
-        this topic will be dispatched to the provided callback.
-
-            callback(message: bytes)
-
-        Updates the internal dictionary topic_callback so that the input
-        callback will be called when a message is received for the given topic.
-
-        :param str topic: topic to subscribe to
-        :param callable callback: callback on message received
+        :param channel: pika.channel.Channel
+        :param basic_deliver: pika.spec.Basic.Deliver
+        :param _basic_properties: pika.spec.Basic.Properties
+        :param body: bytes
         """
-        # 1. Add callback to be called when event on that topic + routing_key
-        # 2. Request a subscription on the new topic towards the consumer
-        #    connection
-        LOGGER.debug("subscribe")
+        consume_params = (self._consumes[basic_deliver.consumer_tag]
+                              .consume_params)
 
-        self.add_new_consumer(Consumer(callback, exchange=topic))
+        try:
+            consume_params.on_message_callback(body)
+        except Exception as e:
+            LOGGER.critical(f"the on_message_callback for queue: "
+                            f"{consume_params.queue} crashed with error: {e}")
 
-        self._connection.subscribe(Subscription(topic))
+        channel.basic_ack(delivery_tag=basic_deliver.delivery_tag)
 
-    def is_subscribed(self, topic):
+    def on_ready(self):
         """
-        Checks if a subscription has been activated, meaning RMQ has confirmed
-        with a ConsumeOk that a consume is active.
-
-        :param topic: topic to check
-        :return: true if active
+        Connection hook, called when channel opened, meaning RMQConnection is
+        ready for work.
         """
-        result = False
+        LOGGER.info("consumer connection ready")
 
-        for consumer in self._consumers:
-            if not consumer.preserve:  # RPC
-                if consumer.exchange == topic and \
-                   consumer.consumer_tag:  # If not "", ConsumeOK gotten
-                    result = True
-                    break
+        self._ready = True
 
-        LOGGER.debug("is_subscribed {}".format(result))
+        for _key, consume in self._consumes.items():
+            self._handle_consume(consume.consume_params,
+                                 consume.queue_params,
+                                 consume.exchange_params,
+                                 consume.routing_key)
 
-        return result
-
-    def rpc_server(self, queue_name, callback):
+    def on_close(self, permanent=False):
         """
-        Starts an RPC server by issuing a subscribe request to the consumer
-        connection.
+        Connection hook, called when the channel or connection is closed.
 
-        The preserve flag ensures that the raw ConsumedMessage object is
-        forwarded to the supplied callback function.
+        NOTE!
+        This is NOT reported as an error because users are expected to
+        configure their exchanges and queues accordingly to avoid losing data
+        they cannot live without.
 
-        :param queue_name: RPC server request queue name
-        :param callback: callback on message received
+        :param permanent: bool
         """
-        LOGGER.debug("rpc_server")
+        if permanent:
+            LOGGER.critical("consumer connection permanently closed")
+        else:
+            LOGGER.info("consumer connection closed")
 
-        self.add_new_consumer(Consumer(callback,
-                                       preserve=True,
-                                       routing_key=queue_name))
-        self._connection.rpc_server(RPCServer(queue_name))
+        self._ready = False
 
-    def rpc_client(self, queue_name, callback):
+        old_consumer_tags = list()
+        for _key, consume in self._consumes.items():
+            # Remove all consumer tag keys in local dict since these may change
+            # on re-subscription.
+            if consume.consumer_tag is not None:
+                old_consumer_tags.append(consume.consumer_tag)
+            consume.consumer_tag = None
+
+        for consumer_tag in old_consumer_tags:
+            self._consumes.pop(consumer_tag, None)
+
+    def on_error(self):
         """
-        Starts an RPC client by issuing a subscribe request to the consumer
-        connection.
-
-        The preserve flag ensures that the raw ConsumedMessage object is
-        forwarded to the supplied callback function.
-
-        :param queue_name: RPC client response queue name
-        :param callback: callback on message received
+        Connection hook, called when the connection has encountered an error.
         """
-        LOGGER.debug("rpc_client")
+        LOGGER.info("consumer connection error")
 
-        self.add_new_consumer(Consumer(callback,
-                                       preserve=True,
-                                       routing_key=queue_name))
-        self._connection.rpc_client(RPCClient(queue_name))
-
-    def is_rpc_consumer_ready(self, server_name) -> bool:
-        """
-        Checks if the RPC server is ready.
-
-        :return: True if ready
-        """
-        result = False
-
-        for consumer in self._consumers:
-            if consumer.preserve:
-                if consumer.routing_key == server_name and \
-                   consumer.consumer_tag:
-                    result = True
-                    break
-
-        LOGGER.debug(f"is_rpc_consumer_ready {result}")
-
-        return result
-
-    def command_queue(self, queue_name, callback):
-        """
-        Adds a command queue instance to the list of consumers and issues a
-        request to the consumer connection to set up the command queue.
-        """
-        LOGGER.debug(f"command_queue {queue_name}")
-
-        self.add_new_consumer(Consumer(callback, routing_key=queue_name))
-
-        self._connection.command_queue(CommandQueue(queue_name))
+        # Possible errors:
+        # * channel died and will not recover
+        # * callback for operation failed
+        # * declaration with faulty parameters attempted
+        # TODO: Add possibility to signal user that an error has occurred.
